Make ident and parentheses parsers available separately from Base()
When implementing a language with custom operators,
only a part of the Base() is useful, yet you can't
easily get that part without the rest of the features.

An example of these parts are private parse*() functions. While
things like parseString() can be easily re-implemented, the
parseParentheses() and parseIdent() are somewhat more complicated,
yet they're not available for re-using externally.

This change keeps the Base() as is, but allows one to re-use at least
the complex parts of it.